file: Separate splitter and hasher in different packages
