+++ file/hasher/param.go
@@ -1,4 +1,4 @@
-package file
+package hasher
 
 import (
 	"sync"
