+++ file/hasher/util_test.go
@@ -1,4 +1,4 @@
-package file
+package hasher
 
 import "testing"
